requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
redis==4.0.2
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1 